+++ Tests/CSharp/ExpressionTests/ExpressionTests.cs
@@ -1020,22 +1020,17 @@ internal partial class TestClass
 {
     private void TestMethod()
     {
-        object test(object a) => a * 2;
+        object test(object a) => Operators.MultiplyObject(a, 2);
         object test2(object a, object b)
         {
-            if (Conversions.ToBoolean(b > (object)0))
-                return a / b;
+            if (Conversions.ToBoolean(Operators.CompareObjectGreater(b, 0, false)))
+                return Operators.DivideObject(a, b);
             return 0;
         };
-        object test3(object a, object b) => a % b;
+        object test3(object a, object b) => Operators.ModObject(a, b);
         test(3);
     }
-}
-4 target compilation errors:
-CS0019: Operator '*' cannot be applied to operands of type 'object' and 'object'
-CS0019: Operator '>' cannot be applied to operands of type 'object' and 'object'
-CS0019: Operator '/' cannot be applied to operands of type 'object' and 'object'
-CS0019: Operator '%' cannot be applied to operands of type 'object' and 'object'");
+}");
         }
 
         [Fact]